+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.2.0] - 2025-11-04
+
+### Changed
+
+- migrated to `bagit-utils`-library (from `bagit`)
+
+### Added
+
+- added SET-operation (sets target field value regardless original data)
+
+### Fixed
+
+- fixed preservation of element order and comments of significant_properties.xml
+
 ## [1.0.0] - 2025-09-09
 
 ### Changed